requirements.txt
flask
flask-sqlalchemy
flask-login
torch
torchvision
pillow
numpy
werkzeug
python-dotenv
psycopg2-binary
cloudinary
